fix: Remove ability to close modal
If the user accidentally or deliberately closes the modal, then the user can restart the game but cannot advance to the next level. This means that there should either be a next level button in the score panel which only appears when a level is won or the user should not be able to close the modal. The latter option was chosen since there is not much of a need to see what's behind the modal when the level is over since all the relevant information is already displayed in the modal.