fix: Adjust level countdown timings
As a result of one the previous updates (1f5c2d2), the user now has to wait for the previous two cards to flip over before opening a third card. This means that it will take longer to complete a level. Also, commit 8bc218c has changed the number of cards in the last two levels as well. Thus, the timings needed to be adjusted accordingly.