fix: Remove icons that won't render correctly
Icons requiring the class 'fab' instead of 'fas' won't render if the CSS version of Font Awesome is used. They will render correctly if the JS version of Font Awesome is used, but this is also incredibly laggy, resulting in a horrible user experience when playing the game. Thus, it was decided to just remove all icons requiring the 'fab' class. This meant that half of the 'Money' option was removed, so this was renamed to 'Currency' and depicted as an 'easy' option.