fix: Decrease number of levels
There are now only three levels, where the first level starts with 16 cards at the request of the Udacity reviewer. More levels are not being added after 36 cards since this would results in card sizes being too tiny for a person to tap on with their finger on smaller mobile devices.